Solve Session problems
