--- src/ChatUapp.Domain.Shared/Core/Interfaces/IAskMessageService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace ChatUapp.Core.Interfaces
-{
-    public interface IAskMessageService
-    {
-        Task<string> AskAnything(string message);
-    }
-}
