+++ src/ChatUapp.Domain.Shared/Core/Interfaces/MessageServices/IAskMessageService.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+
+namespace ChatUapp.Core.Interfaces.MessageServices;
+
+public interface IAskMessageService
+{
+    Task<string> AskAnything(string message);
+}
