--- src/ChatUapp.Domain/Core/ChatbotManagement/Sessions/IChatSessionRepository.cs
@@ -1,8 +1,0 @@
-﻿using ChatUapp.Core.ChatbotManagement.AggregateRoots;
-using Volo.Abp.Domain.Repositories;
-
-namespace ChatUapp.Core.ChatbotManagement.Sessions;
-
-public interface IChatSessionRepository : IRepository<ChatSession>
-{
-}
