--- src/ChatUapp.Infrastructure/AskMessageService.cs
@@ -1,13 +1,0 @@
-﻿using ChatUapp.Core.Interfaces;
-
-namespace ChatUapp.Infrastructure
-{
-    public class AskMessageService : IAskMessageService
-    {
-        public async Task<string> AskAnything(string message)
-        {
-            await Task.Delay(50);
-            return "This message perfectly executed.";
-        }
-    }
-}
